# docs/openapi.yaml
openapi: 3.1.0
info:
  title: EDT Platform Control Plane API
  description: >
    Control-plane REST API for the Enterprise Design Thinking AI Platform (edt_platform).
    Manages projects, runs (the Design Thinking lifecycle), artifacts, human-approval gates,
    events/streaming, agents, evaluations, and semantic memory queries.
    See docs/12-api-specification.md, docs/03-workflow.md, docs/13-data-model.md,
    docs/14-event-model.md.
  version: 1.0.0
  contact:
    name: EDT Platform
    url: https://edt.example.com
  license:
    name: Proprietary
servers:
  - url: https://api.edt.example.com
    description: Production
security:
  - oidc: [runs:read]
...
paths:
...
  /v1/runs:
    get:
      tags: [runs]
      summary: List runs
      operationId: listRuns
      security: [{ oidc: [runs:read] }]
      parameters:
        - name: project_id
          in: query
          schema: { type: string, format: uuid }
        - name: state
          in: query
          schema: { $ref: '#/components/schemas/RunState' }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Paginated runs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Run' }
  /v1/runs/{run_id}:
    get:
      tags: [runs]
      summary: Get run
      operationId: getRun
      security: [{ oidc: [runs:read] }]
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Run' }
        '404': { $ref: '#/components/responses/Error' }
...
  /v1/runs/{run_id}/artifacts:
    get:
      tags: [artifacts]
      summary: List artifacts of a run
      operationId: listRunArtifacts
      security: [{ oidc: [artifacts:read] }]
      parameters:
        - $ref: '#/components/parameters/RunId'
        - name: type
          in: query
          schema: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Artifact' }
...
  /v1/runs/{run_id}/events:
    get:
      tags: [events]
      summary: List persisted events for a run
      operationId: listRunEvents
      security: [{ oidc: [runs:read] }]
      parameters:
        - $ref: '#/components/parameters/RunId'
        - name: type
          in: query
          schema: { type: string }
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/CloudEvent' }
...
components:
  securitySchemes:
    oidc:
      type: openIdConnect
      openIdConnectUrl: https://id.edt.example.com/realms/edt/.well-known/openid-configuration
  parameters:
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
...
    RunId:
      name: run_id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: Error envelope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
          examples:
            validation:
              value:
                error:
                  type: https://errors.edt/validation
                  code: validation_error
                  message: confidence_threshold must be between 0 and 1
                  status: 422
                  trace_id: 4bf92f3577b34da6a3ce929d0e0e4736
  schemas:
    RunState:
      type: string
      enum: [created, planning, discover, awaiting_approval, define, ideate,
             prototype, validate, generating_deliverables, completed, failed, cancelled]
    Page:
      type: object
      properties:
        next_cursor: { type: [string, "null"] }
        has_more: { type: boolean }
...
    Run:
      type: object
      required: [run_id, project_id, state]
      properties:
        run_id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        state: { $ref: '#/components/schemas/RunState' }
        pending_gate: { type: [string, "null"] }
        confidence_threshold: { type: number }
        budget_usd_cap: { type: [number, "null"] }
        cost_usd_spent: { type: number }
        temporal_workflow_id: { type: [string, "null"] }
        phases:
          type: array
          items: { $ref: '#/components/schemas/Phase' }
        created_at: { type: string, format: date-time }
    Phase:
      type: object
      properties:
        name: { type: string }
        seq: { type: integer }
        status: { type: string, enum: [pending, running, completed, failed] }
        avg_confidence: { type: [number, "null"] }
    Artifact:
      type: object
      required: [id, run_id, type, version]
      properties:
        id: { type: string, format: uuid }
        run_id: { type: string, format: uuid }
        phase: { type: string }
        type: { type: string }
        title: { type: [string, "null"] }
        version: { type: integer }
        producer_agent: { type: string }
        confidence: { type: number, minimum: 0, maximum: 1 }
        provenance: { type: object, additionalProperties: true }
        s3_uri: { type: [string, "null"] }
        status: { type: string, enum: [draft, approved, superseded] }
        content: { type: object, additionalProperties: true }
        created_at: { type: string, format: date-time }
...
    CloudEvent:
      type: object
      required: [specversion, id, source, type, time]
      properties:
        specversion: { type: string, enum: ["1.0"] }
        id: { type: string }
        source: { type: string }
        type: { type: string }
        subject: { type: string }
        time: { type: string, format: date-time }
        datacontenttype: { type: string }
        dataschema: { type: string }
        orgid: { type: string }
        runid: { type: string }
        traceid: { type: string }
        data: { type: object, additionalProperties: true }
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, status]
          properties:
            type: { type: string }
            code: { type: string }
            message: { type: string }
            status: { type: integer }
            trace_id: { type: string }
            details:
              type: array
              items:
                type: object
                additionalProperties: true